+++ lib/rubocop.rb
@@ -104,7 +104,7 @@ require 'rubocop/formatter/simple_text_formatter'
 require 'rubocop/formatter/emacs_style_formatter'
 require 'rubocop/formatter/progress_formatter'
 require 'rubocop/formatter/json_formatter'
-require 'rubocop/formatter/details_formatter'
+require 'rubocop/formatter/clang_style_formatter'
 require 'rubocop/formatter/formatter_set'
 
 require 'rubocop/config'