+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * Introduced formatter feature, enables custom formatted output and multiple outputs.
 * Added progress formatter. (`--format progress`)
 * Added JSON formatter. (`--format json`)
-* Added details formatter showing the offending source code. (`--format details`)
+* Added clang style formatter showing the offending source code. (`--format clang`)
 * Added `-f`/`--format` option to specify formatter.
 * Added `-o`/`--out` option to specify output file for each formatter.
 * Added `-r/--require` option to inject external Ruby code into RuboCop.